Split the tracing JsonLayer into 2 layers
THe WorkersJsonLayer is only used when running in prod, this means the
`JsonFields` extension is only present in prod, but it's now needed both
in prod and development environments as it's how we can get the tracing
fields to pass to DOs through request headers.
